refactor(admin): split money route panels
Move the money route display helpers, row renderers, and async data sections into route-local modules so page.tsx owns only layout composition and suspense boundaries.

Verified with pnpm exec prettier --check apps/admin/src/app/money/page.tsx apps/admin/src/app/money/money-display.tsx apps/admin/src/app/money/money-rows.tsx apps/admin/src/app/money/money-sections.tsx; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; git diff --check -- apps/admin/src/app/money/page.tsx apps/admin/src/app/money/money-display.tsx apps/admin/src/app/money/money-rows.tsx apps/admin/src/app/money/money-sections.tsx; pnpm validate; pnpm dlx knip --no-exit-code.
